// src/common/audit/system-audit/system-audit.enum.ts
export enum SystemAction {
  // User Actions
  USER_CREATED = 'USER_CREATED',
  USER_UPDATED = 'USER_UPDATED',
  USER_ROLE_ASSIGNED = 'USER_ROLE_ASSIGNED',
  USER_DELETED = 'USER_DELETED',

  // Role Actions
  ROLE_CREATED = 'ROLE_CREATED',
  ROLE_UPDATED = 'ROLE_UPDATED',
  ROLE_DELETED = 'ROLE_DELETED',
  ROLE_PERMISSIONS_ASSIGNED = 'ROLE_PERMISSIONS_ASSIGNED',
  ROLE_PERMISSIONS_REMOVED = 'ROLE_PERMISSIONS_REMOVED',

  // Permission Actions
  PERMISSION_CREATED = 'PERMISSION_CREATED',
  PERMISSION_UPDATED = 'PERMISSION_UPDATED',
  PERMISSION_DELETED = 'PERMISSION_DELETED',
}